+++ CHANGELOG.md
@@ -16,6 +16,14 @@
 - Corrected CHANGELOG.MD to report that issue with InterfaceAlias matching
   on Adapter description rather than Adapter Name was released in 5.7.0.0
   rather than 5.6.0.0 - See [Issue #315](https://github.com/PowerShell/xNetworking/issues/315).
+- MSFT_WaitForNetworkTeam:
+  - Added a new resource to set the wait for a network team to become 'Up'.
+- MSFT_NetworkTeam:
+  - Improved detection of environmemt for running network team integration
+    tests.
+- MSFT_NetworkTeamInterface:
+  - Improved detection of environmemt for running network team integration
+    tests.
 
 ## 5.7.0.0